requirements.txt
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=0.24.0
Pillow>=9.2.0
streamlit>=1.5.0
tqdm>=4.62.0
plotly
